refactor: NumberGeneratingStrategy의 패키지 변경
- NumberGeneratingStrategy의 패키지 domain -> utils
- LottoTest 주석 추가
